requirements.txt
Django>=4.2,<6.0
python-dotenv>=1.0.0
celery>=5.3
redis>=5.0
django-ratelimit>=4.1.0
qrcode[pil]>=7.4
Pillow>=10.0
python-pptx>=0.6.23
openpyxl>=3.1
gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.2
